[NR_PHY_COMMON]: Move LLR functions to nr_compute_llr.c
The simple LLR functions (nr_qpsk_llr, nr_16qam_llr, nr_64qam_llr,
nr_256qam_llr) were in nr_phy_common.c while all other LLR logic
resides in nr_compute_llr.c. Move them to consolidate LLR code.
Update nr_compute_llr.h with explicit declarations.
Fix test_llr.cpp to include nr_compute_llr.h directly.

Change types to c16_t from opaque int32_t * in LLR ref functions.
Update nr_16qam_llr_ref, nr_64qam_llr_ref, nr_256qam_llr_ref to use
const c16_t * for ch_mag parameters, removing manual int16_t * casts
and accessing .r/.i fields directly.
