+++ openair1/PHY/nr_phy_common/inc/nr_compute_llr.h
@@ -5,7 +5,17 @@
 #ifndef __NR_COMPUTE_LLR__H__
 #define __NR_COMPUTE_LLR__H__
 
-#include "PHY/impl_defs_top.h"
+#include "platform_types.h"
+
+void nr_qpsk_llr(const c16_t *rxdataF_comp, int16_t *llr, uint32_t nb_re);
+void nr_16qam_llr(const c16_t *rxdataF_comp, const c16_t *ch_mag_in, int16_t *llr, uint32_t nb_re);
+void nr_64qam_llr(const c16_t *rxdataF_comp, const c16_t *ch_mag, const c16_t *ch_mag2, int16_t *llr, uint32_t nb_re);
+void nr_256qam_llr(const c16_t *rxdataF_comp,
+                   const c16_t *ch_mag,
+                   const c16_t *ch_mag2,
+                   const c16_t *ch_mag3,
+                   int16_t *llr,
+                   uint32_t nb_re);
 
 void nr_compute_llr(c16_t *rxdataF_comp,
                     c16_t *ch_mag,
@@ -45,6 +55,7 @@ void nr_compute_ML_llr(c16_t *rxdataF_comp0,
                        uint32_t nb_re,
                        uint8_t mod_order);
 
+#ifndef __cplusplus
 uint8_t nr_mmse_2layers(c16_t **rxdataF_comp,
                         uint32_t buffer_length,
                         uint32_t pdsch_buf_size_max,
@@ -60,5 +71,22 @@ uint8_t nr_mmse_2layers(c16_t **rxdataF_comp,
                         unsigned char symbol,
                         int length,
                         uint32_t noise_var);
+#else
+uint8_t nr_mmse_2layers(c16_t **rxdataF_comp,
+                        uint32_t buffer_length,
+                        uint32_t pdsch_buf_size_max,
+                        int nb_rx_ant,
+                        int nb_layers,
+                        c16_t **ch_mag,
+                        c16_t **ch_magb,
+                        c16_t **ch_magc,
+                        c16_t ***ch_estimates_ext,
+                        unsigned short nb_rb,
+                        unsigned char mod_order,
+                        int shift,
+                        unsigned char symbol,
+                        int length,
+                        uint32_t noise_var);
+#endif
 
 #endif /* __NR_COMPUTE_LLR__H__ */
